version 0.3 Сделал базовую логику для ProductService, UserInfoService, ProductController, UserInfoController. Подправил мапперы, теперь они являются бинами. Переделал тип данных у колонок weight и price.

+++ src/main/java/com/pavcore/plumbing/dto/requestTo/ProductRequestTO.java
@@ -16,7 +16,7 @@ public class ProductRequestTO {
     private double weight;
     private String material;
     private String color;
-    private int articleNumber;
+    private long articleNumber;
     private String description;
 
 }
